Remove duplicates by 'interview week' and 'pid'
Rarely, an individual could complete two surveys
(from two separate periods), within a day of each other.
If these days fell in the same week, the derived "interview
week" would be identical for the two survey periods. For
example, we'd have two entries for "week 22". This change
removes duplicates, such that the dataset is unique by PID and
interview week.

This isn't ideal. These are not duplicates, in the sense that
the two entries (in a single week) are from separate survey
periods. So we should think of a better solution here.